Completing Room-Service
Removal of unused files
