+++ .env
@@ -6,7 +6,14 @@ MONGO_DB='db'
 
 # APP
 EXPRESS_PORT=3000
-PRE_SHARED_KEY='password' # TODO: Replace this with key for Collab service
 
-# APP - Matching Service
-QUEUE_EXPIRY=30 * 1000
+# Room Service
+RS_MONGO_USER=''
+RS_MONGO_PASS=''
+RS_MONGO_PORT=27017
+RS_MONGO_DB='roomservice'
+RS_ROOM_EXPIRY=600000
+RS_EXPRESS_PORT=3000
+
+SERVICE_USER_PORT=''
+SERVICE_QUESTION_PORT=9001
