+++ roomApi/src/dataStructs/config.ts
@@ -4,12 +4,7 @@
 
 /** Represents the app's configs. */
 export default class Config {
-  /** Variable names that are found in environment  */
-  private static readonly envVarRedisHost: string = 'REDIS_HOST';
-  private static readonly envVarRedisPort: string = 'REDIS_PORT';
-  private static readonly envVarRedisPassword: string = 'REDIS_PASSWORD';
-  private static readonly envVarRedisExpiry: string = 'QUEUE_EXPIRY';
-
+  // Variable names that are found in environment
   private static readonly envVarMongoHost: string = 'MONGO_HOST';
   private static readonly envVarMongoPort: string = 'MONGO_PORT';
   private static readonly envVarMongoUser: string = 'MONGO_USER';
@@ -20,8 +15,7 @@ export default class Config {
 
   private static readonly envVarExpressPort: string = 'EXPRESS_PORT';
 
-  private static readonly envVarPSK: string = 'PRE_SHARED_KEY';
-
+  // SERVICES
   private static readonly envUserServiceHost: string = 'SERVICE_USER_HOST';
   private static readonly envUserServicePort: string = 'SERVICE_USER_PORT';
 
@@ -33,11 +27,6 @@ export default class Config {
   private static instance: Config | undefined;
 
   /** Copies from Environment and save into these variable names. */
-  public readonly redisHost: string;
-  public readonly redisPort: number;
-  public readonly redisPass: string;
-  public readonly redisExpiry: number;
-
   public readonly mongoHost: string;
   public readonly mongoPort: number;
   public readonly mongoUser: string;
@@ -45,9 +34,7 @@ export default class Config {
   public readonly mongoDB: string;
 
   public readonly mongoRoomExpiry: number;
-
   public readonly expressPort: number;
-  public readonly psk: string | undefined;
 
   public readonly userServiceURI: string | undefined;
   public readonly questionServiceURI: string | undefined;
@@ -60,33 +47,20 @@ export default class Config {
    * @param env - Environment variables.
    */
   private constructor(env: NodeJS.ProcessEnv = process.env) {
-    // Redis
-    this.redisHost =
-      Config._parseString(env[Config.envVarRedisHost]) ?? '127.0.0.1';
-    this.redisPort = Config._parseInt(env[Config.envVarRedisPort]) ?? 6379;
-    this.redisPass = Config._parseString(env[Config.envVarRedisPassword]) ?? '';
-    this.redisExpiry = Config._parseInt(env[Config.envVarRedisExpiry]) ?? 3600;
-
     // Mongo
     this.mongoHost =
       Config._parseString(env[Config.envVarMongoHost]) ?? '127.0.0.1';
-    this.mongoPort = Config._parseInt(env[Config.envVarMongoPort]) ?? 27018;
+    this.mongoPort = Config._parseInt(env[Config.envVarMongoPort]) ?? 27017;
     this.mongoUser = Config._parseString(env[Config.envVarMongoUser]) ?? '';
     this.mongoPass = Config._parseString(env[Config.envVarMongoPass]) ?? '';
-    this.mongoDB = Config._parseString(env[Config.envVarMongoDB]) ?? 'db';
+    this.mongoDB =
+      Config._parseString(env[Config.envVarMongoDB]) ?? 'roomservice';
     this.mongoRoomExpiry =
-      Config._parseInt(env[Config.envVarMongoRoomExpiry]) ?? 30 * 1000; //10 * 60 * 1000;
+      Config._parseInt(env[Config.envVarMongoRoomExpiry]) ?? 1 * 60 * 1000;
+
+    console.log(this.mongoRoomExpiry);
 
     this.expressPort = Config._parseInt(env[Config.envVarExpressPort]) ?? 3000;
-
-    // PSK
-    let temp = Config._parseString(env[Config.envVarPSK]) ?? 'secret-key';
-    if (temp == undefined) {
-      // can't set value here, need throw error.
-      console.log('PSK not initialized!!');
-    } else {
-      this.psk = temp;
-    }
 
     // public readonly userServiceURI: string | undefined;
     // public readonly questionServiceURI: string | undefined;
