+++ run.bat
@@ -1,9 +1,14 @@
 @echo off
-cd ./api
+cd ./api/matchingAPI
 call npm run build
 :: npm run start:dev
 
 call docker build -t matching_service .
+
+
+cd ./api/matchingAPI
+call npm run build
+call docker build -t room_service .
 
 cd ..
 docker-compose up
